Fix error permission and error "no such file or directory"

+++ scripts/user-patches.sh
@@ -184,12 +184,14 @@ function do_patch() {
   fi
 
   # Remove old scripts
-  find "${DMS_CONFIG}/webapi" -mindepth 1 -maxdepth 1 -type d -exec rm -rf {} +
-  find "${DMS_CONFIG}/webapi" -mindepth 1 -maxdepth 1 -type f ! -name ".env" -delete
+  if [ -d "${DMS_CONFIG}/webapi" ]; then
+    find "${DMS_CONFIG}/webapi" -mindepth 1 -maxdepth 1 -type d -exec rm -rf {} +
+    find "${DMS_CONFIG}/webapi" -mindepth 1 -maxdepth 1 -type f ! -name ".env" -delete
+  fi
 
   # Unzip the file
-  unzip docker-mailserver-webapi.zip -d "${DMS_CONFIG}/webapi"
-  rm -rf docker-mailserver-webapi.zip
+  unzip docker-mailserver-webapi.zip -d "${DMS_CONFIG}/webapi" && rm -rf docker-mailserver-webapi.zip
+  chmod -R +x "${DMS_CONFIG}/webapi/scripts"
 
   # Check .env file
   detect_env
